no more warnings,
put animation get list in animation routes
